JSON 데이터 구조체 구분
